Put schema-related code into own .cpp file
